mjfpl/requirements.txt
#packages for development
pytest==6.0.1
flake8==3.9.2
mypy==0.910

#package for app
requests==2.24.0
SQLAlchemy==1.3.19
flask_sqlalchemy
BeautifulSoup4==4.10.0
flask-cors==3.0.10my
flask_restful